Add a static index.d.ts and a reference in package
To make dependent packages recognise the types, we need to add
a declaration that there are indeed types to look for. This commit
adds an `index.d.ts` file in the root that contains references to the
.d.ts files generated in a previous commit in this branch. It also adds
a reference to that file in `package.json`, which is The Correct Way,
even though `index.d.ts` should be sufficient.

However! This requires a follow up commit that generates the
`index.d.ts` file in the build script, rather than leaving it hardcoded.
The current file is good for now, but should be removed and generated at
build.

- Add `types` field in `package.json`
- Add hardcoded `index.d.ts`